+++ src/contexts/PhysicalAssets/domain/physical-asset-service.interface.ts
@@ -1,0 +1,11 @@
+import { IService } from "../../Shared/domain";
+import {
+	PhysicalAsset,
+	PhysicalAssetPrimitives,
+} from "./physical-asset.entity";
+
+export interface IPhysicalAssetService extends IService<
+	string,
+	PhysicalAsset,
+	PhysicalAssetPrimitives
+> {}